+++ Configuration/TCA/Overrides/tt_content.php
@@ -1,8 +1,29 @@
 <?php
 defined('TYPO3') || die();
+use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
+use TYPO3\CMS\Extbase\Utility\ExtensionUtility;
 
-\TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerPlugin(
+$ctypeKey = ExtensionUtility::registerPlugin(
     'tooltip',
     'Replace',
-    'Add Tooltips'
+    'LLL:EXT:tooltip/Resources/Private/Language/locallang_db.xlf:plugin.replace.title'
 );
+
+$ctypeKey =  ExtensionUtility::registerPlugin(
+    'tooltip',
+    'Glossar',
+    'LLL:EXT:tooltip/Resources/Private/Language/locallang_db.xlf:plugin.glossar.title'
+);
+ExtensionManagementUtility::addToAllTCAtypes(
+  'tt_content',
+  '--div--;Configuration,pi_flexform,',
+  $ctypeKey,
+  'after:subheader',
+);
+
+ExtensionManagementUtility::addPiFlexFormValue(
+  '*',
+  'FILE:EXT:tooltip/Configuration/FlexForms/Glossar.xml',
+  $ctypeKey,
+);
+